Generic function for extracting the predictive partial log-likelihood

Generic function for extracting th predictive partial log-likelihood from a fitted survival model.

Usage

PLL(object, newdata, newtime, newstatus, ...)

Arguments

object

fitted model of class class.

newdata

n_new*p matrix of covariates.

newtime

n_new-vector of censored survival times.

newstatus

n_new-vector of event status, coded with 0 for censoring and 1, if an event occurred.

...

additional arguments, for example complexity value, if necessary.

Details

The predictive partial log-likelihood measures the prediction performance of each model fitted in a boostrap sample, using the data not in this sample. Multiplying by (-2) leads to a deviance-like measure, which means that small values indicate good prediction performance.

peperr requires function PLL.class in case of survival response, for each model fit of class class. At the time, PLL.CoxBoost is available.

Value

Vector of length n_new